+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * UX Consistency: Update HTML Page Title display when third pane (detail record) displays. Refs UINV-525.
 * Align the `finance.fund` interface version (`3.0`). Refs UINV-545.
+* Use Save & close button label  stripes-component translation key. Refs UINV-540.
 
 ## [6.0.3](https://github.com/folio-org/ui-invoice/tree/v6.0.3) (2024-04-19)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v6.0.2...v6.0.3)